9.1. Moved server to api folder.
